+++ src/install/hashtopolis.sql
@@ -1133,7 +1133,7 @@ ALTER TABLE `Hash`
   ADD PRIMARY KEY (`hashId`),
   ADD KEY `hashlistId` (`hashlistId`),
   ADD KEY `chunkId` (`chunkId`),
-  ADD KEY `isCraced` (`isCracked`),
+  ADD KEY `isCracked` (`isCracked`),
   ADD KEY `hash` (`hash`);
 
 --